Refactor buffer analysis
