+++ lib/node/index.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var nodes = {
-    Buffer: require('./nodes/buffer'),
+    Buffer: require('./nodes/buffer/buffer'),
     Moran: require('./nodes/moran/moran'),
     PointInPolygon: require('./nodes/point-in-polygon/point-in-polygon'),
     PopulationInArea: require('./nodes/population-in-area/population-in-area'),
