--- lib/node/nodes/buffer.js
@@ -1,26 +1,0 @@
-'use strict';
-
-var Node = require('../node');
-
-var TYPE = 'buffer';
-var PARAMS = {
-    source: Node.PARAM.NODE,
-    radius: Node.PARAM.NUMBER
-};
-
-var Buffer = Node.create(TYPE, PARAMS);
-
-module.exports = Buffer;
-module.exports.TYPE = TYPE;
-module.exports.PARAMS = PARAMS;
-
-Buffer.prototype.sql = function() {
-    var _columns = [
-        'ST_Buffer(the_geom::geography, ' + this.radius + ')::geometry the_geom'
-    ].concat(this.source.getColumns(true));
-    return [
-        'SELECT',
-        _columns.join(','),
-        'FROM (' + this.source.getQuery() + ') _camshaft_buffer'
-    ].join('\n');
-};
