+++ lib/node/nodes/population-in-area/population-in-area-query-builder.js
@@ -9,5 +9,7 @@ dot.templateSettings.strip = false;
 var populationInAreaTemplate = dot.template(populationInAreaQuery);
 
 module.exports = function buildQuery(it) {
+    it.columns = it.columns.join(', ');
+
     return populationInAreaTemplate(it);
 };
