+++ lib/node/nodes/population-in-area/population-in-area.js
@@ -21,5 +21,4 @@ PopulationInArea.prototype.sql = function() {
         columns: this.source.getColumns(),
         final_column: this.final_column
     });
-
 };
